requirements.txt
django>=5.2,<6.0
django-cors-headers>=4.9.0
djangorestframework>=3.16.1
mssql-django>=1.6
pyodbc>=5.1.0
pymongo>=4.4.0
